Update rating in cache without calling server
Since we only want to rate songs at this point in time, it was opted to have a method specifically for that purpose.

When caching, CachedDataKey.Song requires an `api_object.Song`, however it was getting a `models.Song` object
 which triggered an endless loop of trying to update the song artist and the album.
The decision was therefore made to use a caching key specific to updating song ratings.